Add common resolver
